DBConverter/requirement.txt
PyMySQL==1.1.0
python-dotenv==0.21.1
cryptography>=41.0.0